+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 * @chemoish emulated HTMLTrackElement to enable track load events ([view](https://github.com/videojs/video.js/pull/2804))
 * @gkatsev added nullcheck for cues in updateForTrack. Fixes #2870 ([view](https://github.com/videojs/video.js/pull/2896))
 * @gkatsev added ability to release next tag from master ([view](https://github.com/videojs/video.js/pull/2894))
+* @gkatsev added chg- and github- release for next releases ([view](https://github.com/videojs/video.js/pull/2899))
 
 --------------------